Add files via upload
This Project aims to do overall testing like functional testing, GUI testing on Demo tour website to check quality of application using selenium web driver. It helps to improve quality of website and saves time for manually doing this testing. 
In this project our main focus is to test the registration and flight booking module. A tourism website is usually seeking to appeal to a broad audience and often has a wealth of information to share with that audience.Finding the best design for your requirements means picking you battles .